+++ doc/api/_generate_api.rst
@@ -11,6 +11,6 @@ To include a new package/module, list it below and include it in api.index.rst
    :toctree: ./
 
    gmt
-   gmt.clib
    gmt.exceptions
-
+   gmt.clib
+   gmt.helpers
