+++ doc/api/index.rst
@@ -3,26 +3,7 @@
 API Reference
 =============
 
-.. raw:: html
-
-    <div class="api-module">
-
 .. include:: gmt.rst
-
-.. raw:: html
-
-    </div>
-    <div class="api-module">
-
-.. include:: gmt.clib.rst
-
-.. raw:: html
-
-    </div>
-    <div class="api-module">
-
 .. include:: gmt.exceptions.rst
-
-.. raw:: html
-
-    </div>
+.. include:: gmt.clib.rst
+.. include:: gmt.helpers.rst
