+++ gmt/figure.py
@@ -2,10 +2,7 @@
 Define the Figure class that handles all plotting.
 """
 import os
-import sys
-import subprocess
-import webbrowser
-from tempfile import NamedTemporaryFile, TemporaryDirectory
+from tempfile import TemporaryDirectory
 import base64
 
 try:
@@ -15,9 +12,9 @@ except ImportError:
 
 from .clib import LibGMT
 from .base_plotting import BasePlotting
-from .utils import build_arg_string
-from .decorators import fmt_docstring, use_alias, kwargs_to_strings
 from .exceptions import GMTError
+from .helpers import build_arg_string, fmt_docstring, use_alias, \
+    kwargs_to_strings, launch_external_viewer, unique_name
 
 
 def figure(name):
@@ -43,51 +40,6 @@ def figure(name):
     fmt = '-'
     with LibGMT() as lib:
         lib.call_module('figure', '{} {}'.format(name, fmt))
-
-
-def unique_name():
-    """
-    Generate a unique name for a figure.
-
-    Need a unique name for each figure, otherwise GMT will plot everything
-    on the same figure instead of creating a new one.
-
-    Returns
-    -------
-    name : str
-        A unique name generated by ``tempfile.NamedTemporaryFile``
-
-    """
-    # Use the tempfile module to generate a unique file name.
-    with NamedTemporaryFile(prefix='gmt-python-') as tmpfile:
-        return os.path.split(tmpfile.name)[-1]
-
-
-def launch_external_viewer(fname):
-    """
-    Open a file in an external viewer program.
-
-    Uses the ``xdg-open`` command on Linux, the ``open`` command on macOS, and
-    the default web browser on other systems.
-
-    Parameters
-    ----------
-    fname : str
-        The file name of the file (preferably a full path).
-
-    """
-    # Redirect stdout and stderr to devnull so that the terminal isn't filled
-    # with noise
-    run_args = dict(stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-
-    # Open the file with the default viewer.
-    # Fall back to the browser if can't recognize the operating system.
-    if sys.platform.startswith('linux'):
-        subprocess.run(['xdg-open', fname], **run_args)
-    elif sys.platform == 'darwin':  # Darwin is macOS
-        subprocess.run(['open', fname], **run_args)
-    else:
-        webbrowser.open_new_tab('file://{}'.format(fname))
 
 
 class Figure(BasePlotting):
