+++ gmt/helpers/utils.py
@@ -1,12 +1,12 @@
 """
 Utilities and common tasks for wrapping the GMT modules.
 """
-import os
+import sys
+import subprocess
+import webbrowser
 from contextlib import contextmanager
-from tempfile import NamedTemporaryFile
-import numpy as np
 
-from .exceptions import GMTError
+from ..exceptions import GMTError
 
 
 def data_kind(data, x, y):
@@ -185,83 +185,28 @@ def is_nonstr_iter(value):
     return bool(not isinstance(value, str) and is_iterable)
 
 
-class GMTTempFile():
+def launch_external_viewer(fname):
     """
-    Context manager for creating closed temporary files.
+    Open a file in an external viewer program.
 
-    This class does not return a file-like object. So, you can't do
-    ``for line in GMTTempFile()``, for example, or pass it to things that
-    need file objects.
+    Uses the ``xdg-open`` command on Linux, the ``open`` command on macOS, and
+    the default web browser on other systems.
 
     Parameters
     ----------
-    prefix : str
-        The temporary file name begins with the prefix.
-    suffix : str
-        The temporary file name ends with the suffix.
+    fname : str
+        The file name of the file (preferably a full path).
 
-    Examples
-    --------
-    >>> import numpy as np
-    >>> with GMTTempFile() as tmpfile:
-    ...     # write data to temporary file
-    ...     x = y = z = np.arange(0, 3, 1)
-    ...     np.savetxt(tmpfile.name, (x, y, z), fmt='%.1f')
-    ...     lines = tmpfile.read()
-    ...     print(lines)
-    ...     nx, ny, nz = tmpfile.loadtxt(unpack=True, dtype=float)
-    ...     print(nx, ny, nz)
-    0.0 1.0 2.0
-    0.0 1.0 2.0
-    0.0 1.0 2.0
-    <BLANKLINE>
-    [0. 0. 0.] [1. 1. 1.] [2. 2. 2.]
     """
-    def __init__(self, prefix="gmt-python-", suffix=".txt"):
-        args = dict(prefix=prefix, suffix=suffix, delete=False)
-        with NamedTemporaryFile(**args) as tmpfile:
-            self.name = tmpfile.name
+    # Redirect stdout and stderr to devnull so that the terminal isn't filled
+    # with noise
+    run_args = dict(stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args):
-        if os.path.exists(self.name):
-            os.remove(self.name)
-
-    def read(self, keep_tabs=False):
-        """
-        Read the entire contents of the file as a Unicode string.
-
-        Parameters
-        ----------
-        keep_tabs : bool
-            If False, replace the tabs that GMT uses with spaces.
-
-        Returns
-        -------
-        content : str
-            Content of the temporary file as a Unicode string.
-        """
-        with open(self.name) as tmpfile:
-            content = tmpfile.read()
-            if not keep_tabs:
-                content = content.replace('\t', ' ')
-            return content
-
-    def loadtxt(self, **kwargs):
-        """
-        Load data from the temporary file using numpy.loadtxt.
-
-        Parameters
-        ----------
-        kwargs : dict
-            Any keyword arguments that can be passed to numpy.loadtxt.
-
-        Returns
-        -------
-        ndarray
-            Data read from the text file.
-
-        """
-        return np.loadtxt(self.name, **kwargs)
+    # Open the file with the default viewer.
+    # Fall back to the browser if can't recognize the operating system.
+    if sys.platform.startswith('linux'):
+        subprocess.run(['xdg-open', fname], **run_args)
+    elif sys.platform == 'darwin':  # Darwin is macOS
+        subprocess.run(['open', fname], **run_args)
+    else:
+        webbrowser.open_new_tab('file://{}'.format(fname))
