+++ gmt/modules.py
@@ -1,11 +1,8 @@
 """
 Non-plot GMT modules.
 """
-
 from .clib import LibGMT
-from .utils import build_arg_string
-from .decorators import fmt_docstring
-from .utils import GMTTempFile
+from .helpers import build_arg_string, fmt_docstring, GMTTempFile
 
 
 @fmt_docstring
