+++ gmt/tests/test_figure.py
@@ -7,14 +7,6 @@ import os
 import pytest
 
 from .. import Figure
-from ..figure import unique_name
-
-
-def test_figure_unique_name():
-    "Make sure the figure names start with gmt-python- and are really unique"
-    names = [unique_name() for i in range(100)]
-    assert all([name.startswith('gmt-python-') for name in names])
-    assert len(names) == len(set(names))
 
 
 def test_figure_savefig_exists():
